Fix ScrollView layout and vertical bounce behavior
When UIRefreshControl is added to UIScrollView, it follows the
UIScrollView contentInset. UIRefreshControl's frame's width is
always the scroll view's frame's width. This causes the
refresh control to look off center. Subtract the scroll view's
contentInset.left to account for this.

Adding UIRefreshControl to a UIScrollView with disabled
`alwaysBounceVertical` whose content size is less than the
scroll view's height will not allow pulling on the scroll view
to refresh. If the refresh behavior is not `disabled` then
set `alwaysBounceVertical` to `true`.